+++ gradio/interface.py
@@ -14,7 +14,7 @@ from typing import TYPE_CHECKING, Any, Callable
 
 from gradio_client.documentation import document, set_documentation_group
 
-from gradio import Examples, external, interpretation, utils
+from gradio import Examples, analytics, external, interpretation, utils
 from gradio.blocks import Blocks
 from gradio.components import (
     Button,
@@ -362,7 +362,7 @@ class Interface(Blocks):
         self.local_url = None
 
         self.favicon_path = None
-        utils.version_check()
+        analytics.version_check()
         Interface.instances.add(self)
 
         param_names = inspect.getfullargspec(self.fn)[0]
